Refactor Concourse.java
Move the implementation of the methods to a separate class
