+++ concourse-driver-java/src/main/java/com/cinchapi/concourse/Concourse.java
@@ -15,50 +15,21 @@
  */
 package com.cinchapi.concourse;
 
-import java.nio.ByteBuffer;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Set;
-import java.util.concurrent.Callable;
-
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.NotThreadSafe;
-
-import org.apache.thrift.TException;
-import org.apache.thrift.protocol.TBinaryProtocol;
-import org.apache.thrift.protocol.TProtocol;
-import org.apache.thrift.transport.TSocket;
-import org.apache.thrift.transport.TTransport;
-import org.apache.thrift.transport.TTransportException;
 
 import com.cinchapi.concourse.annotate.Incubating;
 import com.cinchapi.concourse.config.ConcourseClientPreferences;
 import com.cinchapi.concourse.lang.BuildableState;
 import com.cinchapi.concourse.lang.Criteria;
-import com.cinchapi.concourse.lang.Language;
-import com.cinchapi.concourse.security.ClientSecurity;
-import com.cinchapi.concourse.thrift.AccessToken;
-import com.cinchapi.concourse.thrift.ComplexTObject;
-import com.cinchapi.concourse.thrift.ConcourseService;
 import com.cinchapi.concourse.thrift.Diff;
 import com.cinchapi.concourse.thrift.Operator;
-import com.cinchapi.concourse.thrift.SecurityException;
-import com.cinchapi.concourse.thrift.TObject;
-import com.cinchapi.concourse.thrift.TransactionToken;
-import com.cinchapi.concourse.time.Time;
-import com.cinchapi.concourse.util.ByteBuffers;
-import com.cinchapi.concourse.util.Collections;
-import com.cinchapi.concourse.util.Conversions;
 import com.cinchapi.concourse.util.Convert;
-import com.cinchapi.concourse.util.PrettyLinkedHashMap;
-import com.cinchapi.concourse.util.PrettyLinkedTableMap;
-import com.cinchapi.concourse.util.Transformers;
-import com.google.common.base.Preconditions;
-import com.google.common.base.Throwables;
 import com.google.common.collect.Sets;
-import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
 
 /**
@@ -107,7 +78,7 @@ public abstract class Concourse implements AutoCloseable {
      * @return the handle
      */
     public static Concourse connect() {
-        return new Client();
+        return new ConcourseThriftDriver();
     }
 
     /**
@@ -120,7 +91,7 @@ public abstract class Concourse implements AutoCloseable {
      * @return the handle
      */
     public static Concourse connect(String environment) {
-        return new Client(environment);
+        return new ConcourseThriftDriver(environment);
     }
 
     /**
@@ -135,7 +106,7 @@ public abstract class Concourse implements AutoCloseable {
      */
     public static Concourse connect(String host, int port, String username,
             String password) {
-        return new Client(host, port, username, password);
+        return new ConcourseThriftDriver(host, port, username, password);
     }
 
     /**
@@ -152,7 +123,7 @@ public abstract class Concourse implements AutoCloseable {
      */
     public static Concourse connect(String host, int port, String username,
             String password, String environment) {
-        return new Client(host, port, username, password, environment);
+        return new ConcourseThriftDriver(host, port, username, password, environment);
     }
 
     /**
@@ -3143,3176 +3114,4 @@ public abstract class Concourse implements AutoCloseable {
      * @return a copy of this connection handle
      */
     protected abstract Concourse copyConnection();
-
-    /**
-     * The implementation of the {@link Concourse} interface that establishes a
-     * connection with the remote server and handles communication. This class
-     * is a more user friendly wrapper around a Thrift
-     * {@link ConcourseService.Client}.
-     * 
-     * @author Jeff Nelson
-     */
-    private final static class Client extends Concourse {
-
-        private static String ENVIRONMENT;
-        private static String PASSWORD;
-        private static String SERVER_HOST;
-        private static int SERVER_PORT;
-        private static String USERNAME;
-        static {
-            // If there is a concourse_client.prefs file located in the working
-            // directory, parse it and use its values as defaults.
-            ConcourseClientPreferences config;
-            try {
-                config = ConcourseClientPreferences
-                        .open("concourse_client.prefs");
-            }
-            catch (Exception e) {
-                config = null;
-            }
-            SERVER_HOST = "localhost";
-            SERVER_PORT = 1717;
-            USERNAME = "admin";
-            PASSWORD = "admin";
-            ENVIRONMENT = "";
-            if(config != null) {
-                SERVER_HOST = config.getString("host", SERVER_HOST);
-                SERVER_PORT = config.getInt("port", SERVER_PORT);
-                USERNAME = config.getString("username", USERNAME);
-                PASSWORD = config.getString("password", PASSWORD);
-                ENVIRONMENT = config.getString("environment", ENVIRONMENT);
-            }
-        }
-
-        /**
-         * The Thrift client that actually handles all RPC communication.
-         */
-        private final ConcourseService.Client client;
-
-        /**
-         * The client keeps a copy of its {@link AccessToken} and passes it to
-         * the server for each remote procedure call. The client will
-         * re-authenticate when necessary using the username/password read from
-         * the prefs file.
-         */
-        private AccessToken creds = null;
-
-        /**
-         * The environment to which the client is connected.
-         */
-        private final String environment;
-
-        /**
-         * The host of the connection.
-         */
-        private final String host;
-
-        /**
-         * An encrypted copy of the password passed to the constructor.
-         */
-        private final ByteBuffer password;
-
-        /**
-         * The port of the connection.
-         */
-        private final int port;
-
-        /**
-         * Whenever the client starts a Transaction, it keeps a
-         * {@link TransactionToken} so that the server can stage the changes in
-         * the appropriate place.
-         */
-        private TransactionToken transaction = null;
-
-        /**
-         * An encrypted copy of the username passed to the constructor.
-         */
-        private final ByteBuffer username;
-
-        /**
-         * Create a new Client connection to the environment of the Concourse
-         * Server described in {@code concourse_client.prefs} (or the default
-         * environment and server if the prefs file does not exist) and return a
-         * handler to facilitate database interaction.
-         */
-        public Client() {
-            this(ENVIRONMENT);
-        }
-
-        /**
-         * Create a new Client connection to the specified {@code environment}
-         * of the Concourse Server described in {@code concourse_client.prefs}
-         * (or the default server if the prefs file does not exist) and return a
-         * handler to facilitate database interaction.
-         * 
-         * @param environment
-         */
-        public Client(String environment) {
-            this(SERVER_HOST, SERVER_PORT, USERNAME, PASSWORD, environment);
-        }
-
-        /**
-         * Create a new Client connection to the default environment of the
-         * specified Concourse Server and return a handler to facilitate
-         * database interaction.
-         * 
-         * @param host
-         * @param port
-         * @param username
-         * @param password
-         */
-        public Client(String host, int port, String username, String password) {
-            this(host, port, username, password, "");
-        }
-
-        /**
-         * Create a new Client connection to the specified {@code environment}
-         * of the specified Concourse Server and return a handler to facilitate
-         * database interaction.
-         * 
-         * @param host
-         * @param port
-         * @param username
-         * @param password
-         * @param environment
-         */
-        public Client(String host, int port, String username, String password,
-                String environment) {
-            this.host = host;
-            this.port = port;
-            this.username = ClientSecurity.encrypt(username);
-            this.password = ClientSecurity.encrypt(password);
-            this.environment = environment;
-            final TTransport transport = new TSocket(host, port);
-            try {
-                transport.open();
-                TProtocol protocol = new TBinaryProtocol(transport);
-                client = new ConcourseService.Client(protocol);
-                authenticate();
-                Runtime.getRuntime().addShutdownHook(new Thread("shutdown") {
-
-                    @Override
-                    public void run() {
-                        if(transaction != null && transport.isOpen()) {
-                            abort();
-                            transport.close();
-                        }
-                    }
-
-                });
-            }
-            catch (TTransportException e) {
-                throw new RuntimeException(
-                        "Could not connect to the Concourse Server at " + host
-                                + ":" + port);
-            }
-        }
-
-        @Override
-        public void abort() {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    if(transaction != null) {
-                        final TransactionToken token = transaction;
-                        transaction = null;
-                        client.abort(creds, token, environment);
-                    }
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> long add(final String key, final T value) {
-            return execute(new Callable<Long>() {
-
-                @Override
-                public Long call() throws Exception {
-                    return client.addKeyValue(key, Convert.javaToThrift(value),
-                            creds, transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Boolean> add(final String key, final T value,
-                final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Boolean>>() {
-
-                @Override
-                public Map<Long, Boolean> call() throws Exception {
-                    Map<Long, Boolean> raw = client.addKeyValueRecords(key,
-                            Convert.javaToThrift(value),
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                    Map<Long, Boolean> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", "Successful");
-                    for (long record : records) {
-                        pretty.put(record, raw.get(record));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> boolean add(final String key, final T value,
-                final long record) {
-            return execute(new Callable<Boolean>() {
-
-                @Override
-                public Boolean call() throws Exception {
-                    return client.addKeyValueRecord(key,
-                            Convert.javaToThrift(value), record, creds,
-                            transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Timestamp, String> audit(final long record) {
-            return execute(new Callable<Map<Timestamp, String>>() {
-
-                @Override
-                public Map<Timestamp, String> call() throws Exception {
-                    Map<Long, String> audit = client.auditRecord(record, creds,
-                            transaction, environment);
-                    return ((PrettyLinkedHashMap<Timestamp, String>) Transformers
-                            .transformMap(audit,
-                                    Conversions.timestampToMicros()))
-                            .setKeyName("DateTime").setValueName("Revision");
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Timestamp, String> audit(final long record,
-                final Timestamp start) {
-            return execute(new Callable<Map<Timestamp, String>>() {
-
-                @Override
-                public Map<Timestamp, String> call() throws Exception {
-                    Map<Long, String> audit;
-                    if(start.isString()) {
-                        audit = client.auditRecordStartstr(record,
-                                start.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        audit = client.auditRecordStart(record,
-                                start.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    return ((PrettyLinkedHashMap<Timestamp, String>) Transformers
-                            .transformMap(audit,
-                                    Conversions.timestampToMicros()))
-                            .setKeyName("DateTime").setValueName("Revision");
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Timestamp, String> audit(final long record,
-                final Timestamp start, final Timestamp end) {
-            return execute(new Callable<Map<Timestamp, String>>() {
-
-                @Override
-                public Map<Timestamp, String> call() throws Exception {
-                    Map<Long, String> audit;
-                    if(start.isString()) {
-                        audit = client.auditRecordStartstrEndstr(record,
-                                start.toString(), end.toString(), creds,
-                                transaction, environment);
-                    }
-                    else {
-                        audit = client.auditRecordStartEnd(record,
-                                start.getMicros(), end.getMicros(), creds,
-                                transaction, environment);
-                    }
-                    return ((PrettyLinkedHashMap<Timestamp, String>) Transformers
-                            .transformMap(audit,
-                                    Conversions.timestampToMicros()))
-                            .setKeyName("DateTime").setValueName("Revision");
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Timestamp, String> audit(final String key, final long record) {
-            return execute(new Callable<Map<Timestamp, String>>() {
-
-                @Override
-                public Map<Timestamp, String> call() throws Exception {
-                    Map<Long, String> audit = client.auditKeyRecord(key,
-                            record, creds, transaction, environment);
-                    return ((PrettyLinkedHashMap<Timestamp, String>) Transformers
-                            .transformMap(audit,
-                                    Conversions.timestampToMicros()))
-                            .setKeyName("DateTime").setValueName("Revision");
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Timestamp, String> audit(final String key,
-                final long record, final Timestamp start) {
-            return execute(new Callable<Map<Timestamp, String>>() {
-
-                @Override
-                public Map<Timestamp, String> call() throws Exception {
-                    Map<Long, String> audit;
-                    if(start.isString()) {
-                        audit = client.auditKeyRecordStartstr(key, record,
-                                start.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        audit = client.auditKeyRecordStart(key, record,
-                                start.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    return ((PrettyLinkedHashMap<Timestamp, String>) Transformers
-                            .transformMap(audit,
-                                    Conversions.timestampToMicros()))
-                            .setKeyName("DateTime").setValueName("Revision");
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Timestamp, String> audit(final String key,
-                final long record, final Timestamp start, final Timestamp end) {
-            return execute(new Callable<Map<Timestamp, String>>() {
-
-                @Override
-                public Map<Timestamp, String> call() throws Exception {
-                    Map<Long, String> audit;
-                    if(start.isString()) {
-                        audit = client.auditKeyRecordStartstrEndstr(key,
-                                record, start.toString(), end.toString(),
-                                creds, transaction, environment);
-                    }
-                    else {
-                        audit = client.auditKeyRecordStartEnd(key, record,
-                                start.getMicros(), end.getMicros(), creds,
-                                transaction, environment);
-                    }
-                    return ((PrettyLinkedHashMap<Timestamp, String>) Transformers
-                            .transformMap(audit,
-                                    Conversions.timestampToMicros()))
-                            .setKeyName("DateTime").setValueName("Revision");
-                }
-
-            });
-        }
-
-        @Override
-        public Map<String, Map<Object, Set<Long>>> browse(
-                final Collection<String> keys) {
-            return execute(new Callable<Map<String, Map<Object, Set<Long>>>>() {
-
-                @Override
-                public Map<String, Map<Object, Set<Long>>> call()
-                        throws Exception {
-                    Map<String, Map<TObject, Set<Long>>> raw = client
-                            .browseKeys(Collections.toList(keys), creds,
-                                    transaction, environment);
-                    Map<String, Map<Object, Set<Long>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Key");
-                    for (Entry<String, Map<TObject, Set<Long>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.thriftToJava(),
-                                        Conversions.<Long> none()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Map<String, Map<Object, Set<Long>>> browse(
-                final Collection<String> keys, final Timestamp timestamp) {
-            return execute(new Callable<Map<String, Map<Object, Set<Long>>>>() {
-
-                @Override
-                public Map<String, Map<Object, Set<Long>>> call()
-                        throws Exception {
-                    Map<String, Map<TObject, Set<Long>>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.browseKeysTimestr(
-                                Collections.toList(keys), timestamp.toString(),
-                                creds, transaction, environment);
-                    }
-                    else {
-                        raw = client.browseKeysTime(Collections.toList(keys),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<String, Map<Object, Set<Long>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Key");
-                    for (Entry<String, Map<TObject, Set<Long>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.thriftToJava(),
-                                        Conversions.<Long> none()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Map<Object, Set<Long>> browse(final String key) {
-            return execute(new Callable<Map<Object, Set<Long>>>() {
-
-                @Override
-                public Map<Object, Set<Long>> call() throws Exception {
-                    Map<TObject, Set<Long>> raw = client.browseKey(key, creds,
-                            transaction, environment);
-                    Map<Object, Set<Long>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap(key, "Records");
-                    for (Entry<TObject, Set<Long>> entry : raw.entrySet()) {
-                        pretty.put(Convert.thriftToJava(entry.getKey()),
-                                entry.getValue());
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Map<Object, Set<Long>> browse(final String key,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Object, Set<Long>>>() {
-
-                @Override
-                public Map<Object, Set<Long>> call() throws Exception {
-                    Map<TObject, Set<Long>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.browseKeyTimestr(key,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.browseKeyTime(key, timestamp.getMicros(),
-                                creds, transaction, environment);
-                    }
-                    Map<Object, Set<Long>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap(key, "Records");
-                    for (Entry<TObject, Set<Long>> entry : raw.entrySet()) {
-                        pretty.put(Convert.thriftToJava(entry.getKey()),
-                                entry.getValue());
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Map<Timestamp, Set<Object>> chronologize(final String key,
-                final long record) {
-            return execute(new Callable<Map<Timestamp, Set<Object>>>() {
-
-                @Override
-                public Map<Timestamp, Set<Object>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw = client.chronologizeKeyRecord(
-                            key, record, creds, transaction, environment);
-                    Map<Timestamp, Set<Object>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("DateTime", "Values");
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                Timestamp.fromMicros(entry.getKey()),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.thriftToJava()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Timestamp, Set<Object>> chronologize(final String key,
-                final long record, final Timestamp start) {
-            Preconditions.checkArgument(start.getMicros() <= Time.now(),
-                    "Start of range cannot be greater than the present");
-            return execute(new Callable<Map<Timestamp, Set<Object>>>() {
-
-                @Override
-                public Map<Timestamp, Set<Object>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw;
-                    if(start.isString()) {
-                        raw = client.chronologizeKeyRecordStartstr(key, record,
-                                start.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.chronologizeKeyRecordStart(key, record,
-                                start.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Timestamp, Set<Object>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("DateTime", "Values");
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                Timestamp.fromMicros(entry.getKey()),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.thriftToJava()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Map<Timestamp, Set<Object>> chronologize(final String key,
-                final long record, final Timestamp start, final Timestamp end) {
-            return execute(new Callable<Map<Timestamp, Set<Object>>>() {
-
-                @Override
-                public Map<Timestamp, Set<Object>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw;
-                    if(start.isString()) {
-                        raw = client.chronologizeKeyRecordStartstrEndstr(key,
-                                record, start.toString(), end.toString(),
-                                creds, transaction, environment);
-                    }
-                    else {
-                        raw = client.chronologizeKeyRecordStartEnd(key, record,
-                                start.getMicros(), end.getMicros(), creds,
-                                transaction, environment);
-                    }
-                    Map<Timestamp, Set<Object>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("DateTime", "Values");
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                Timestamp.fromMicros(entry.getKey()),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.thriftToJava()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public void clear(final Collection<Long> records) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.clearRecords(Collections.toLongList(records), creds,
-                            transaction, environment);
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public void clear(final Collection<String> keys,
-                final Collection<Long> records) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.clearKeysRecords(Collections.toList(keys),
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public void clear(final Collection<String> keys, final long record) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.clearKeysRecord(Collections.toList(keys), record,
-                            creds, transaction, environment);
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public void clear(final long record) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.clearRecord(record, creds, transaction, environment);
-                    return null;
-                }
-
-            });
-
-        }
-
-        @Override
-        public void clear(final String key, final Collection<Long> records) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.clearKeyRecords(key,
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public void clear(final String key, final long record) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.clearKeyRecord(key, record, creds, transaction,
-                            environment);
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public boolean commit() {
-            return execute(new Callable<Boolean>() {
-
-                @Override
-                public Boolean call() throws Exception {
-                    final TransactionToken token = transaction;
-                    transaction = null;
-                    return token != null ? client.commit(creds, token,
-                            environment) : false;
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Long, Set<String>> describe(final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Set<String>>>() {
-
-                @Override
-                public Map<Long, Set<String>> call() throws Exception {
-                    Map<Long, Set<String>> raw = client.describeRecords(
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                    Map<Long, Set<String>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", "Keys");
-                    for (Entry<Long, Set<String>> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(), entry.getValue());
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Map<Long, Set<String>> describe(final Collection<Long> records,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Set<String>>>() {
-
-                @Override
-                public Map<Long, Set<String>> call() throws Exception {
-                    Map<Long, Set<String>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.describeRecordsTimestr(
-                                Collections.toLongList(records),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.describeRecordsTime(
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Set<String>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", "Keys");
-                    for (Entry<Long, Set<String>> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(), entry.getValue());
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Set<String> describe(final long record) {
-            return execute(new Callable<Set<String>>() {
-
-                @Override
-                public Set<String> call() throws Exception {
-                    Set<String> result = client.describeRecord(record, creds,
-                            transaction, environment);
-                    return result;
-                }
-            });
-        }
-
-        @Override
-        public Set<String> describe(final long record, final Timestamp timestamp) {
-            return execute(new Callable<Set<String>>() {
-
-                @Override
-                public Set<String> call() throws Exception {
-                    if(timestamp.isString()) {
-                        return client.describeRecordTimestr(record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        return client.describeRecordTime(record,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                }
-            });
-        }
-
-        @Override
-        public <T> Map<String, Map<Diff, Set<T>>> diff(final long record,
-                final Timestamp start) {
-            return execute(new Callable<Map<String, Map<Diff, Set<T>>>>() {
-
-                @Override
-                public Map<String, Map<Diff, Set<T>>> call() throws Exception {
-                    Map<String, Map<Diff, Set<TObject>>> raw;
-                    if(start.isString()) {
-                        raw = client.diffRecordStartstr(record,
-                                start.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.diffRecordStart(record, start.getMicros(),
-                                creds, transaction, environment);
-                    }
-                    PrettyLinkedTableMap<String, Diff, Set<T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap();
-                    pretty.setRowName("Key");
-                    for (Entry<String, Map<Diff, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<Diff> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public <T> Map<String, Map<Diff, Set<T>>> diff(final long record,
-                final Timestamp start, final Timestamp end) {
-            return execute(new Callable<Map<String, Map<Diff, Set<T>>>>() {
-
-                @Override
-                public Map<String, Map<Diff, Set<T>>> call() throws Exception {
-                    Map<String, Map<Diff, Set<TObject>>> raw;
-                    if(start.isString()) {
-                        raw = client.diffRecordStartstrEndstr(record,
-                                start.toString(), end.toString(), creds,
-                                transaction, environment);
-                    }
-                    else {
-                        raw = client.diffRecordStartEnd(record,
-                                start.getMicros(), end.getMicros(), creds,
-                                transaction, environment);
-                    }
-                    PrettyLinkedTableMap<String, Diff, Set<T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap();
-                    pretty.setRowName("Key");
-                    for (Entry<String, Map<Diff, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<Diff> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public <T> Map<Diff, Set<T>> diff(final String key, final long record,
-                final Timestamp start) {
-            return execute(new Callable<Map<Diff, Set<T>>>() {
-
-                @Override
-                public Map<Diff, Set<T>> call() throws Exception {
-                    Map<Diff, Set<TObject>> raw;
-                    if(start.isString()) {
-                        raw = client.diffKeyRecordStartstr(key, record,
-                                start.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.diffKeyRecordStart(key, record,
-                                start.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Diff, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Operation", "Value");
-                    for (Entry<Diff, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public <T> Map<Diff, Set<T>> diff(final String key, final long record,
-                final Timestamp start, final Timestamp end) {
-            return execute(new Callable<Map<Diff, Set<T>>>() {
-
-                @Override
-                public Map<Diff, Set<T>> call() throws Exception {
-                    Map<Diff, Set<TObject>> raw;
-                    if(start.isString()) {
-                        raw = client.diffKeyRecordStartstrEndstr(key, record,
-                                start.toString(), end.toString(), creds,
-                                transaction, environment);
-                    }
-                    else {
-                        raw = client.diffKeyRecordStartEnd(key, record,
-                                start.getMicros(), end.getMicros(), creds,
-                                transaction, environment);
-                    }
-                    Map<Diff, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Operation", "Value");
-                    for (Entry<Diff, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        @SuppressWarnings("unchecked")
-        public <T> Map<T, Map<Diff, Set<Long>>> diff(final String key,
-                final Timestamp start) {
-            return execute(new Callable<Map<T, Map<Diff, Set<Long>>>>() {
-
-                @Override
-                public Map<T, Map<Diff, Set<Long>>> call() throws Exception {
-                    Map<TObject, Map<Diff, Set<Long>>> raw;
-                    if(start.isString()) {
-                        raw = client.diffKeyStartstr(key, start.toString(),
-                                creds, transaction, environment);
-                    }
-                    else {
-                        raw = client.diffKeyStart(key, start.getMicros(),
-                                creds, transaction, environment);
-                    }
-                    PrettyLinkedTableMap<T, Diff, Set<Long>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap();
-                    pretty.setRowName("Value");
-                    for (Entry<TObject, Map<Diff, Set<Long>>> entry : raw
-                            .entrySet()) {
-                        pretty.put((T) Convert.thriftToJava(entry.getKey()),
-                                entry.getValue());
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        @SuppressWarnings("unchecked")
-        public <T> Map<T, Map<Diff, Set<Long>>> diff(final String key,
-                final Timestamp start, final Timestamp end) {
-            return execute(new Callable<Map<T, Map<Diff, Set<Long>>>>() {
-
-                @Override
-                public Map<T, Map<Diff, Set<Long>>> call() throws Exception {
-                    Map<TObject, Map<Diff, Set<Long>>> raw;
-                    if(start.isString()) {
-                        raw = client.diffKeyStartstrEndstr(key,
-                                start.toString(), end.toString(), creds,
-                                transaction, environment);
-                    }
-                    else {
-                        raw = client.diffKeyStartEnd(key, start.getMicros(),
-                                end.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    PrettyLinkedTableMap<T, Diff, Set<Long>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap();
-                    pretty.setRowName("Value");
-                    for (Entry<TObject, Map<Diff, Set<Long>>> entry : raw
-                            .entrySet()) {
-                        pretty.put((T) Convert.thriftToJava(entry.getKey()),
-                                entry.getValue());
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public void exit() {
-            try {
-                client.logout(creds, environment);
-                client.getInputProtocol().getTransport().close();
-                client.getOutputProtocol().getTransport().close();
-            }
-            catch (com.cinchapi.concourse.thrift.SecurityException
-                    | TTransportException e) {
-                // Handle corner case where the client is existing because of
-                // (or after the occurrence of) a password change, which means
-                // it can't perform a traditional logout. Its worth nothing that
-                // we're okay with this scenario because a password change will
-                // delete all previously issued tokens.
-            }
-            catch (Exception e) {
-                throw Throwables.propagate(e);
-            }
-        }
-
-        @Override
-        public Set<Long> find(final Criteria criteria) {
-            return execute(new Callable<Set<Long>>() {
-
-                @Override
-                public Set<Long> call() throws Exception {
-                    return client.findCriteria(
-                            Language.translateToThriftCriteria(criteria),
-                            creds, transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public Set<Long> find(Object object) {
-            if(object instanceof BuildableState) {
-                return find(((BuildableState) object).build());
-            }
-            else {
-                throw new IllegalArgumentException(object
-                        + " is not a valid argument for the find method");
-            }
-        }
-
-        @Override
-        public Set<Long> find(final String ccl) {
-            return execute(new Callable<Set<Long>>() {
-
-                @Override
-                public Set<Long> call() throws Exception {
-                    return client.findCcl(ccl, creds, transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public Set<Long> find(String key, Object value) {
-            return find0(key, Operator.EQUALS, value);
-        }
-
-        @Override
-        public Set<Long> find(String key, Object value, Timestamp timestamp) {
-            return find0(key, Operator.EQUALS, value, timestamp);
-        }
-
-        @Override
-        public Set<Long> find(String key, Operator operator, Object value) {
-            return find0(key, operator, value);
-        }
-
-        @Override
-        public Set<Long> find(String key, Operator operator, Object value,
-                Object value2) {
-            return find0(key, operator, value, value2);
-        }
-
-        @Override
-        public Set<Long> find(String key, Operator operator, Object value,
-                Object value2, Timestamp timestamp) {
-            return find0(timestamp, key, operator, value, value2);
-        }
-
-        @Override
-        public Set<Long> find(final String key, final Operator operator,
-                final Object value, final Timestamp timestamp) {
-            return find0(timestamp, key, operator, value);
-        }
-
-        @Override
-        public Set<Long> find(String key, String operator, Object value) {
-            return find0(key, operator, value);
-        }
-
-        @Override
-        public Set<Long> find(String key, String operator, Object value,
-                final Object value2) {
-            return find0(key, operator, value, value2);
-        }
-
-        @Override
-        public Set<Long> find(String key, String operator, Object value,
-                final Object value2, Timestamp timestamp) {
-            return find0(timestamp, key, operator, value, value2);
-        }
-
-        @Override
-        public Set<Long> find(String key, String operator, Object value,
-                Timestamp timestamp) {
-            return find0(timestamp, key, operator, value);
-        }
-
-        @Override
-        public <T> long findOrAdd(final String key, final T value) {
-            return execute(new Callable<Long>() {
-
-                @Override
-                public Long call() throws Exception {
-                    return client.findOrAddKeyValue(key,
-                            Convert.javaToThrift(value), creds, transaction,
-                            environment);
-                }
-
-            });
-        }
-
-        @Override
-        public long findOrInsert(final Criteria criteria, final String json) {
-            return execute(new Callable<Long>() {
-
-                @Override
-                public Long call() throws Exception {
-                    return client.findOrInsertCriteriaJson(
-                            Language.translateToThriftCriteria(criteria), json,
-                            creds, transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public long findOrInsert(final String ccl, final String json) {
-            return execute(new Callable<Long>() {
-
-                @Override
-                public Long call() throws Exception {
-                    return client.findOrInsertCclJson(ccl, json, creds,
-                            transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final Collection<String> keys,
-                final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw = client
-                            .getKeysRecords(Collections.toList(keys),
-                                    Collections.toLongList(records), creds,
-                                    transaction, environment);
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final Collection<String> keys,
-                final Collection<Long> records, final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getKeysRecordsTimestr(
-                                Collections.toList(keys),
-                                Collections.toLongList(records),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getKeysRecordsTime(
-                                Collections.toList(keys),
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final Collection<String> keys,
-                final Criteria criteria) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw = client
-                            .getKeysCriteria(Collections.toList(keys), Language
-                                    .translateToThriftCriteria(criteria),
-                                    creds, transaction, environment);
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final Collection<String> keys,
-                final Criteria criteria, final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getKeysCriteriaTimestr(
-                                Collections.toList(keys),
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getKeysCriteriaTime(
-                                Collections.toList(keys),
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<String, T> get(final Collection<String> keys,
-                final long record) {
-            return execute(new Callable<Map<String, T>>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public Map<String, T> call() throws Exception {
-                    Map<String, TObject> raw = client.getKeysRecord(
-                            Collections.toList(keys), record, creds,
-                            transaction, environment);
-                    Map<String, T> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Key", "Value");
-                    for (Entry<String, TObject> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(),
-                                (T) Convert.thriftToJava(entry.getValue()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<String, T> get(final Collection<String> keys,
-                final long record, final Timestamp timestamp) {
-            return execute(new Callable<Map<String, T>>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public Map<String, T> call() throws Exception {
-                    Map<String, TObject> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getKeysRecordTimestr(
-                                Collections.toList(keys), record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getKeysRecordTime(
-                                Collections.toList(keys), record,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<String, T> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Key", "Value");
-                    for (Entry<String, TObject> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(),
-                                (T) Convert.thriftToJava(entry.getValue()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-                Object criteria) {
-            if(criteria instanceof BuildableState) {
-                return get(keys, ((BuildableState) criteria).build());
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the get method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-                Object criteria, Timestamp timestamp) {
-            if(criteria instanceof BuildableState) {
-                return get(keys, ((BuildableState) criteria).build(), timestamp);
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the get method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final Collection<String> keys,
-                final String ccl) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw = client.getKeysCcl(
-                            Collections.toList(keys), ccl, creds, transaction,
-                            environment);
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final Collection<String> keys,
-                final String ccl, final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getKeysCclTimestr(
-                                Collections.toList(keys), ccl,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getKeysCclTime(Collections.toList(keys),
-                                ccl, timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final Criteria criteria) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw = client.getCriteria(
-                            Language.translateToThriftCriteria(criteria),
-                            creds, transaction, environment);
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final Criteria criteria,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getCriteriaTimestr(
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getCriteriaTime(
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Object criteria) {
-            if(criteria instanceof BuildableState) {
-                return get(((BuildableState) criteria).build());
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the get method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Object criteria,
-                Timestamp timestamp) {
-            if(criteria instanceof BuildableState) {
-                return get(((BuildableState) criteria).build(), timestamp);
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the get method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final String ccl) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw = client.getCcl(ccl,
-                            creds, transaction, environment);
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, T> get(final String key,
-                final Collection<Long> records) {
-            return execute(new Callable<Map<Long, T>>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public Map<Long, T> call() throws Exception {
-                    Map<Long, TObject> raw = client.getKeyRecords(key,
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                    Map<Long, T> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, TObject> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(),
-                                (T) Convert.thriftToJava(entry.getValue()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, T> get(final String key,
-                final Collection<Long> records, final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, T>>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public Map<Long, T> call() throws Exception {
-                    Map<Long, TObject> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getKeyRecordsTimestr(key,
-                                Collections.toLongList(records),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getKeyRecordsTime(key,
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, T> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, TObject> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(),
-                                (T) Convert.thriftToJava(entry.getValue()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, T> get(final String key, final Criteria criteria) {
-            return execute(new Callable<Map<Long, T>>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public Map<Long, T> call() throws Exception {
-                    Map<Long, TObject> raw = client.getKeyCriteria(key,
-                            Language.translateToThriftCriteria(criteria),
-                            creds, transaction, environment);
-                    Map<Long, T> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, TObject> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(),
-                                (T) Convert.thriftToJava(entry.getValue()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, T> get(final String key, final Criteria criteria,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, T>>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public Map<Long, T> call() throws Exception {
-                    Map<Long, TObject> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getKeyCriteriaTimestr(key,
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getKeyCriteriaTime(key,
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, T> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, TObject> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(),
-                                (T) Convert.thriftToJava(entry.getValue()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        @Nullable
-        public <T> T get(final String key, final long record) {
-            return execute(new Callable<T>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public T call() throws Exception {
-                    TObject raw = client.getKeyRecord(key, record, creds,
-                            transaction, environment);
-                    return raw == TObject.NULL ? null : (T) Convert
-                            .thriftToJava(raw);
-
-                }
-
-            });
-        }
-
-        @Override
-        @Nullable
-        public <T> T get(final String key, final long record,
-                final Timestamp timestamp) {
-            return execute(new Callable<T>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public T call() throws Exception {
-                    TObject raw;
-                    if(timestamp.isString()) {
-                        raw = client.getKeyRecordTimestr(key, record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getKeyRecordTime(key, record,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    return raw == TObject.NULL ? null : (T) Convert
-                            .thriftToJava(raw);
-
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, T> get(String key, Object criteria) {
-            if(criteria instanceof BuildableState) {
-                return get(key, ((BuildableState) criteria).build());
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the get method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, T> get(String key, Object criteria,
-                Timestamp timestamp) {
-            if(criteria instanceof BuildableState) {
-                return get(key, ((BuildableState) criteria).build(), timestamp);
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the get method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, T> get(final String key, final String ccl) {
-            return execute(new Callable<Map<Long, T>>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public Map<Long, T> call() throws Exception {
-                    Map<Long, TObject> raw = client.getKeyCcl(key, ccl, creds,
-                            transaction, environment);
-                    Map<Long, T> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, TObject> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(),
-                                (T) Convert.thriftToJava(entry.getValue()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, T> get(final String key, final String ccl,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, T>>() {
-
-                @SuppressWarnings("unchecked")
-                @Override
-                public Map<Long, T> call() throws Exception {
-                    Map<Long, TObject> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getKeyCclTimestr(key, ccl,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.getKeyCclTime(key, ccl,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, T> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, TObject> entry : raw.entrySet()) {
-                        pretty.put(entry.getKey(),
-                                (T) Convert.thriftToJava(entry.getValue()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(final String ccl,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, T>>>() {
-
-                @Override
-                public Map<Long, Map<String, T>> call() throws Exception {
-                    Map<Long, Map<String, TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.getCclTimestr(ccl, timestamp.toString(),
-                                creds, transaction, environment);
-                    }
-                    else {
-                        raw = client.getCclTime(ccl, timestamp.getMicros(),
-                                creds, transaction, environment);
-                    }
-                    Map<Long, Map<String, T>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, TObject>> entry : raw
-                            .entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformMapValues(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public String getServerEnvironment() {
-            return execute(new Callable<String>() {
-
-                @Override
-                public String call() throws Exception {
-                    return client.getServerEnvironment(creds, transaction,
-                            environment);
-                }
-
-            });
-        }
-
-        @Override
-        public String getServerVersion() {
-            return execute(new Callable<String>() {
-
-                @Override
-                public String call() throws Exception {
-                    return client.getServerVersion();
-                }
-
-            });
-        }
-
-        @Override
-        public Set<Long> insert(final String json) {
-            return execute(new Callable<Set<Long>>() {
-
-                @Override
-                public Set<Long> call() throws Exception {
-                    return client.insertJson(json, creds, transaction,
-                            environment);
-                }
-
-            });
-
-        }
-
-        @Override
-        public Map<Long, Boolean> insert(final String json,
-                final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Boolean>>() {
-
-                @Override
-                public Map<Long, Boolean> call() throws Exception {
-                    return client.insertJsonRecords(json,
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public boolean insert(final String json, final long record) {
-            return execute(new Callable<Boolean>() {
-
-                @Override
-                public Boolean call() throws Exception {
-                    return client.insertJsonRecord(json, record, creds,
-                            transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public Set<Long> inventory() {
-            return execute(new Callable<Set<Long>>() {
-
-                @Override
-                public Set<Long> call() throws Exception {
-                    return client.inventory(creds, transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public <T> T invokePlugin(final String id, final String method,
-                final Object... args) {
-            return execute(new Callable<T>() {
-
-                @Override
-                public T call() throws Exception {
-                    List<ComplexTObject> params = Lists
-                            .newArrayListWithCapacity(args.length);
-                    for (Object arg : args) {
-                        params.add(ComplexTObject.fromJavaObject(arg));
-                    }
-                    ComplexTObject result = client.invokePlugin(id, method,
-                            params, creds, transaction, environment);
-                    return result.getJavaObject();
-                }
-
-            });
-        }
-
-        @Override
-        public String jsonify(Collection<Long> records) {
-            return jsonify(records, true);
-        }
-
-        @Override
-        public String jsonify(final Collection<Long> records,
-                final boolean identifier) {
-            return execute(new Callable<String>() {
-
-                @Override
-                public String call() throws Exception {
-                    return client.jsonifyRecords(
-                            Collections.toLongList(records), identifier, creds,
-                            transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public String jsonify(Collection<Long> records, Timestamp timestamp) {
-            return jsonify(records, timestamp, true);
-        }
-
-        @Override
-        public String jsonify(final Collection<Long> records,
-                final Timestamp timestamp, final boolean identifier) {
-            return execute(new Callable<String>() {
-
-                @Override
-                public String call() throws Exception {
-                    if(timestamp.isString()) {
-                        return client.jsonifyRecordsTimestr(
-                                Collections.toLongList(records),
-                                timestamp.toString(), identifier, creds,
-                                transaction, environment);
-                    }
-                    else {
-                        return client.jsonifyRecordsTime(
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), identifier, creds,
-                                transaction, environment);
-                    }
-                }
-
-            });
-        }
-
-        @Override
-        public String jsonify(long record) {
-            return jsonify(Lists.newArrayList(record), true);
-        }
-
-        @Override
-        public String jsonify(long record, boolean identifier) {
-            return jsonify(Lists.newArrayList(record), identifier);
-        }
-
-        @Override
-        public String jsonify(long record, Timestamp timestamp) {
-            return jsonify(Lists.newArrayList(record), timestamp, true);
-        }
-
-        @Override
-        public String jsonify(long record, Timestamp timestamp,
-                boolean identifier) {
-            return jsonify(Lists.newArrayList(record), timestamp, identifier);
-        }
-
-        @Override
-        public Map<Long, Boolean> link(String key,
-                Collection<Long> destinations, long source) {
-            Map<Long, Boolean> result = PrettyLinkedHashMap
-                    .newPrettyLinkedHashMap("Record", "Result");
-            for (long destination : destinations) {
-                result.put(destination, link(key, destination, source));
-            }
-            return result;
-        }
-
-        @Override
-        public boolean link(String key, long destination, long source) {
-            return add(key, Link.to(destination), source);
-        }
-
-        @Override
-        public Map<Long, Boolean> ping(final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Boolean>>() {
-
-                @Override
-                public Map<Long, Boolean> call() throws Exception {
-                    return client.pingRecords(Collections.toLongList(records),
-                            creds, transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public boolean ping(final long record) {
-            return execute(new Callable<Boolean>() {
-
-                @Override
-                public Boolean call() throws Exception {
-                    return client.pingRecord(record, creds, transaction,
-                            environment);
-                }
-
-            });
-        }
-
-        @Override
-        public <T> void reconcile(final String key, final long record,
-                final Collection<T> values) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    Set<TObject> valueSet = Sets
-                            .newHashSetWithExpectedSize(values.size());
-                    for (T value : values) {
-                        valueSet.add(Convert.javaToThrift(value));
-                    }
-                    client.reconcileKeyRecordValues(key, record, valueSet,
-                            creds, transaction, environment);
-                    return null;
-                }
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Boolean> remove(final String key, final T value,
-                final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Boolean>>() {
-
-                @Override
-                public Map<Long, Boolean> call() throws Exception {
-                    Map<Long, Boolean> raw = client.removeKeyValueRecords(key,
-                            Convert.javaToThrift(value),
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                    Map<Long, Boolean> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", "Result");
-                    for (long record : records) {
-                        pretty.put(record, raw.get(record));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> boolean remove(final String key, final T value,
-                final long record) {
-            return execute(new Callable<Boolean>() {
-
-                @Override
-                public Boolean call() throws Exception {
-                    return client.removeKeyValueRecord(key,
-                            Convert.javaToThrift(value), record, creds,
-                            transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public void revert(final Collection<String> keys,
-                final Collection<Long> records, final Timestamp timestamp) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    if(timestamp.isString()) {
-                        client.revertKeysRecordsTimestr(
-                                Collections.toList(keys),
-                                Collections.toLongList(records),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        client.revertKeysRecordsTime(Collections.toList(keys),
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public void revert(final Collection<String> keys, final long record,
-                final Timestamp timestamp) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    if(timestamp.isString()) {
-                        client.revertKeysRecordTimestr(
-                                Collections.toList(keys), record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        client.revertKeysRecordTime(Collections.toList(keys),
-                                record, timestamp.getMicros(), creds,
-                                transaction, environment);
-                    }
-                    return null;
-                }
-
-            });
-
-        }
-
-        @Override
-        public void revert(final String key, final Collection<Long> records,
-                final Timestamp timestamp) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    if(timestamp.isString()) {
-                        client.revertKeyRecordsTimestr(key,
-                                Collections.toLongList(records),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        client.revertKeyRecordsTime(key,
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    return null;
-                }
-
-            });
-
-        }
-
-        @Override
-        public void revert(final String key, final long record,
-                final Timestamp timestamp) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    if(timestamp.isString()) {
-                        client.revertKeyRecordTimestr(key, record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        client.revertKeyRecordTime(key, record,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    return null;
-                }
-
-            });
-
-        }
-
-        @Override
-        public Set<Long> search(final String key, final String query) {
-            return execute(new Callable<Set<Long>>() {
-
-                @Override
-                public Set<Long> call() throws Exception {
-                    return client.search(key, query, creds, transaction,
-                            environment);
-                }
-
-            });
-        }
-
-        @Override
-        public Map<Long, Map<String, Set<Object>>> select(
-                final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Map<String, Set<Object>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<Object>>> call()
-                        throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw = client
-                            .selectRecords(Collections.toLongList(records),
-                                    creds, transaction, environment);
-                    Map<Long, Map<String, Set<Object>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.thriftToJava()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Map<Long, Map<String, Set<Object>>> select(
-                final Collection<Long> records, final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, Set<Object>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<Object>>> call()
-                        throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectRecordsTimestr(
-                                Collections.toLongList(records),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectRecordsTime(
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, Set<Object>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.thriftToJava()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                final Collection<String> keys, final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw = client
-                            .selectKeysRecords(Collections.toList(keys),
-                                    Collections.toLongList(records), creds,
-                                    transaction, environment);
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                final Collection<String> keys, final Collection<Long> records,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectKeysRecordsTimestr(
-                                Collections.toList(keys),
-                                Collections.toLongList(records),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectKeysRecordsTime(
-                                Collections.toList(keys),
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                final Collection<String> keys, final Criteria criteria) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw = client
-                            .selectKeysCriteria(
-                                    Collections.toList(keys),
-                                    Language.translateToThriftCriteria(criteria),
-                                    creds, transaction, environment);
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                final Collection<String> keys, final Criteria criteria,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectKeysCriteriaTimestr(
-                                Collections.toList(keys),
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectKeysCriteriaTime(
-                                Collections.toList(keys),
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<String, Set<T>> select(final Collection<String> keys,
-                final long record) {
-            return execute(new Callable<Map<String, Set<T>>>() {
-
-                @Override
-                public Map<String, Set<T>> call() throws Exception {
-                    Map<String, Set<TObject>> raw = client.selectKeysRecord(
-                            Collections.toList(keys), record, creds,
-                            transaction, environment);
-                    Map<String, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Key", "Values");
-                    for (Entry<String, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<String, Set<T>> select(final Collection<String> keys,
-                final long record, final Timestamp timestamp) {
-            return execute(new Callable<Map<String, Set<T>>>() {
-
-                @Override
-                public Map<String, Set<T>> call() throws Exception {
-                    Map<String, Set<TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectKeysRecordTimestr(
-                                Collections.toList(keys), record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectKeysRecordTime(
-                                Collections.toList(keys), record,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<String, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Key", "Values");
-                    for (Entry<String, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                Collection<String> keys, Object criteria) {
-            if(criteria instanceof BuildableState) {
-                return select(keys, ((BuildableState) criteria).build());
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the select method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                Collection<String> keys, Object criteria, Timestamp timestamp) {
-            if(criteria instanceof BuildableState) {
-                return select(keys, ((BuildableState) criteria).build(),
-                        timestamp);
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the select method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                final Collection<String> keys, final String ccl) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw = client
-                            .selectKeysCcl(Collections.toList(keys), ccl,
-                                    creds, transaction, environment);
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                final Collection<String> keys, final String ccl,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectKeysCclTimestr(
-                                Collections.toList(keys), ccl,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectKeysCclTime(
-                                Collections.toList(keys), ccl,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(final Criteria criteria) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw = client
-                            .selectCriteria(Language
-                                    .translateToThriftCriteria(criteria),
-                                    creds, transaction, environment);
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                final Criteria criteria, final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectCriteriaTimestr(
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectCriteriaTime(
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public Map<String, Set<Object>> select(final long record) {
-            return execute(new Callable<Map<String, Set<Object>>>() {
-
-                @Override
-                public Map<String, Set<Object>> call() throws Exception {
-                    Map<String, Set<TObject>> raw = client.selectRecord(record,
-                            creds, transaction, environment);
-                    Map<String, Set<Object>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Key", "Values");
-                    for (Entry<String, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.thriftToJava()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public Map<String, Set<Object>> select(final long record,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<String, Set<Object>>>() {
-
-                @Override
-                public Map<String, Set<Object>> call() throws Exception {
-                    Map<String, Set<TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectRecordTimestr(record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectRecordTime(record,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<String, Set<Object>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Key", "Values");
-                    for (Entry<String, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.thriftToJava()));
-                    }
-                    return pretty;
-                }
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(Object criteria) {
-            if(criteria instanceof BuildableState) {
-                return select(((BuildableState) criteria).build());
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the get method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(Object criteria,
-                Timestamp timestamp) {
-            if(criteria instanceof BuildableState) {
-                return select(((BuildableState) criteria).build(), timestamp);
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the get method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(final String ccl) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw = client
-                            .selectCcl(ccl, creds, transaction, environment);
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(final String key,
-                final Collection<Long> records) {
-            return execute(new Callable<Map<Long, Set<T>>>() {
-
-                @Override
-                public Map<Long, Set<T>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw = client.selectKeyRecords(key,
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                    Map<Long, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(final String key,
-                final Collection<Long> records, final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Set<T>>>() {
-
-                @Override
-                public Map<Long, Set<T>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectKeyRecordsTimestr(key,
-                                Collections.toLongList(records),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectKeyRecordsTime(key,
-                                Collections.toLongList(records),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(final String key,
-                final Criteria criteria) {
-            return execute(new Callable<Map<Long, Set<T>>>() {
-
-                @Override
-                public Map<Long, Set<T>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw = client.selectKeyCriteria(key,
-                            Language.translateToThriftCriteria(criteria),
-                            creds, transaction, environment);
-                    Map<Long, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(final String key,
-                final Criteria criteria, final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Set<T>>>() {
-
-                @Override
-                public Map<Long, Set<T>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectKeyCriteriaTimestr(key,
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectKeyCriteriaTime(key,
-                                Language.translateToThriftCriteria(criteria),
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Set<T> select(final String key, final long record) {
-            return execute(new Callable<Set<T>>() {
-
-                @Override
-                public Set<T> call() throws Exception {
-                    Set<TObject> values = client.selectKeyRecord(key, record,
-                            creds, transaction, environment);
-                    return Transformers.transformSetLazily(values,
-                            Conversions.<T> thriftToJavaCasted());
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Set<T> select(final String key, final long record,
-                final Timestamp timestamp) {
-            return execute(new Callable<Set<T>>() {
-
-                @Override
-                public Set<T> call() throws Exception {
-                    Set<TObject> values;
-                    if(timestamp.isString()) {
-                        values = client.selectKeyRecordTimestr(key, record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        values = client.selectKeyRecordTime(key, record,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    return Transformers.transformSetLazily(values,
-                            Conversions.<T> thriftToJavaCasted());
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(String key, Object criteria) {
-            if(criteria instanceof BuildableState) {
-                return select(key, ((BuildableState) criteria).build());
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the select method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(String key, Object criteria,
-                Timestamp timestamp) {
-            if(criteria instanceof BuildableState) {
-                return select(key, ((BuildableState) criteria).build(),
-                        timestamp);
-            }
-            else {
-                throw new IllegalArgumentException(criteria
-                        + " is not a valid argument for the select method");
-            }
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(final String key, final String ccl) {
-            return execute(new Callable<Map<Long, Set<T>>>() {
-
-                @Override
-                public Map<Long, Set<T>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw = client.selectKeyCcl(key, ccl,
-                            creds, transaction, environment);
-                    Map<Long, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(final String key, final String ccl,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Set<T>>>() {
-
-                @Override
-                public Map<Long, Set<T>> call() throws Exception {
-                    Map<Long, Set<TObject>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectKeyCclTimestr(key, ccl,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectKeyCclTime(key, ccl,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    Map<Long, Set<T>> pretty = PrettyLinkedHashMap
-                            .newPrettyLinkedHashMap("Record", key);
-                    for (Entry<Long, Set<TObject>> entry : raw.entrySet()) {
-                        pretty.put(
-                                entry.getKey(),
-                                Transformers.transformSetLazily(
-                                        entry.getValue(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(final String ccl,
-                final Timestamp timestamp) {
-            return execute(new Callable<Map<Long, Map<String, Set<T>>>>() {
-
-                @Override
-                public Map<Long, Map<String, Set<T>>> call() throws Exception {
-                    Map<Long, Map<String, Set<TObject>>> raw;
-                    if(timestamp.isString()) {
-                        raw = client.selectCclTimestr(ccl,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        raw = client.selectCclTime(ccl, timestamp.getMicros(),
-                                creds, transaction, environment);
-                    }
-                    Map<Long, Map<String, Set<T>>> pretty = PrettyLinkedTableMap
-                            .newPrettyLinkedTableMap("Record");
-                    for (Entry<Long, Map<String, Set<TObject>>> entry : raw
-                            .entrySet()) {
-                        pretty.put(entry.getKey(), Transformers
-                                .transformMapSet(entry.getValue(),
-                                        Conversions.<String> none(),
-                                        Conversions.<T> thriftToJavaCasted()));
-                    }
-                    return pretty;
-                }
-
-            });
-        }
-
-        @Override
-        public void set(final String key, final Object value,
-                final Collection<Long> records) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.setKeyValueRecords(key, Convert.javaToThrift(value),
-                            Collections.toLongList(records), creds,
-                            transaction, environment);
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public <T> void set(final String key, final T value, final long record) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.setKeyValueRecord(key, Convert.javaToThrift(value),
-                            record, creds, transaction, environment);
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public void stage() throws TransactionException {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    transaction = client.stage(creds, environment);
-                    return null;
-                }
-
-            });
-        }
-
-        @Override
-        public Timestamp time() {
-            return execute(new Callable<Timestamp>() {
-
-                @Override
-                public Timestamp call() throws Exception {
-                    return Timestamp.fromMicros(client.time(creds, transaction,
-                            environment));
-                }
-
-            });
-        }
-
-        @Override
-        public Timestamp time(final String phrase) {
-            return execute(new Callable<Timestamp>() {
-
-                @Override
-                public Timestamp call() throws Exception {
-                    return Timestamp.fromMicros(client.timePhrase(phrase,
-                            creds, transaction, environment));
-                }
-
-            });
-        }
-
-        @Override
-        public String toString() {
-            return "Connected to " + host + ":" + port + " as "
-                    + new String(ClientSecurity.decrypt(username).array());
-        }
-
-        @Override
-        public boolean unlink(String key, long destination, long source) {
-            return remove(key, Link.to(destination), source);
-        }
-
-        @Override
-        public boolean verify(final String key, final Object value,
-                final long record) {
-            return execute(new Callable<Boolean>() {
-
-                @Override
-                public Boolean call() throws Exception {
-                    return client.verifyKeyValueRecord(key,
-                            Convert.javaToThrift(value), record, creds,
-                            transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public boolean verify(final String key, final Object value,
-                final long record, final Timestamp timestamp) {
-            return execute(new Callable<Boolean>() {
-
-                @Override
-                public Boolean call() throws Exception {
-                    if(timestamp.isString()) {
-                        return client.verifyKeyValueRecordTimestr(key,
-                                Convert.javaToThrift(value), record,
-                                timestamp.toString(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        return client.verifyKeyValueRecordTime(key,
-                                Convert.javaToThrift(value), record,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                }
-
-            });
-        }
-
-        @Override
-        public boolean verifyAndSwap(final String key, final Object expected,
-                final long record, final Object replacement) {
-            return execute(new Callable<Boolean>() {
-
-                @Override
-                public Boolean call() throws Exception {
-                    return client.verifyAndSwap(key,
-                            Convert.javaToThrift(expected), record,
-                            Convert.javaToThrift(replacement), creds,
-                            transaction, environment);
-                }
-
-            });
-        }
-
-        @Override
-        public void verifyOrSet(final String key, final Object value,
-                final long record) {
-            execute(new Callable<Void>() {
-
-                @Override
-                public Void call() throws Exception {
-                    client.verifyOrSet(key, Convert.javaToThrift(value),
-                            record, creds, transaction, environment);
-                    return null;
-                }
-
-            });
-        }
-
-        /**
-         * Authenticate the {@link #username} and {@link #password} and populate
-         * {@link #creds} with the appropriate AccessToken.
-         */
-        private void authenticate() {
-            try {
-                creds = client.login(ClientSecurity.decrypt(username),
-                        ClientSecurity.decrypt(password), environment);
-            }
-            catch (TException e) {
-                throw Throwables.propagate(e);
-            }
-        }
-
-        /**
-         * Execute the task defined in {@code callable}. This method contains
-         * retry logic to handle cases when {@code creds} expires and must be
-         * updated.
-         * 
-         * @param callable
-         * @return the task result
-         */
-        private <T> T execute(Callable<T> callable) {
-            try {
-                return callable.call();
-            }
-            catch (SecurityException e) {
-                authenticate();
-                return execute(callable);
-            }
-            catch (com.cinchapi.concourse.thrift.TransactionException e) {
-                throw new TransactionException();
-            }
-            catch (com.cinchapi.concourse.thrift.DuplicateEntryException e) {
-                throw new DuplicateEntryException(e);
-            }
-            catch (com.cinchapi.concourse.thrift.InvalidArgumentException e) {
-                throw new InvalidArgumentException(e);
-            }
-            catch (com.cinchapi.concourse.thrift.ParseException e) {
-                throw new ParseException(e);
-            }
-            catch (Exception e) {
-                throw Throwables.propagate(e);
-            }
-        }
-
-        /**
-         * Perform an old-school/simple find operation where {@code key}
-         * satisfied {@code operation} in relation to the specified
-         * {@code values}.
-         * 
-         * @param key
-         * @param operator
-         * @param values
-         * @return the records that match the criteria.
-         */
-        private Set<Long> find0(final String key, final Object operator,
-                final Object... values) {
-            final List<TObject> tValues = Lists.transform(
-                    Lists.newArrayList(values), Conversions.javaToThrift());
-            return execute(new Callable<Set<Long>>() {
-
-                @Override
-                public Set<Long> call() throws Exception {
-                    if(operator instanceof Operator) {
-                        return client.findKeyOperatorValues(key,
-                                (Operator) operator, tValues, creds,
-                                transaction, environment);
-                    }
-                    else {
-                        return client.findKeyOperatorstrValues(key,
-                                operator.toString(), tValues, creds,
-                                transaction, environment);
-                    }
-
-                }
-
-            });
-        }
-
-        /**
-         * Perform an old-school/simple find operation where {@code key}
-         * satisfied {@code operation} in relation to the specified
-         * {@code values} at {@code timestamp}.
-         * 
-         * @param key
-         * @param operator
-         * @param values
-         * @param timestamp a {@link Timestamp} that represents the historical
-         *            instant to use in the lookup – created from either a
-         *            {@link Timestamp#fromString(String) natural language
-         *            description} of a point in time (i.e. two weeks ago), OR
-         *            the {@link Timestamp#fromMicros(long) number
-         *            of microseconds} since the Unix epoch, OR
-         *            a {@link Timestamp#fromJoda(org.joda.time.DateTime) Joda
-         *            DateTime} object
-         * @return the records that match the criteria.
-         */
-        private Set<Long> find0(final Timestamp timestamp, final String key,
-                final Object operator, final Object... values) {
-            final List<TObject> tValues = Lists.transform(
-                    Lists.newArrayList(values), Conversions.javaToThrift());
-            return execute(new Callable<Set<Long>>() {
-
-                @Override
-                public Set<Long> call() throws Exception {
-                    if(operator instanceof Operator) {
-                        return client.findKeyOperatorValuesTime(key,
-                                (Operator) operator, tValues,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-                    else {
-                        return client.findKeyOperatorstrValuesTime(key,
-                                operator.toString(), tValues,
-                                timestamp.getMicros(), creds, transaction,
-                                environment);
-                    }
-
-                }
-
-            });
-        }
-
-        @Override
-        protected final Concourse copyConnection() {
-            return new Client(host, port, ByteBuffers.getString(ClientSecurity
-                    .decrypt(username)), ByteBuffers.getString(ClientSecurity
-                    .decrypt(password)), environment);
-        }
-
-    }
 }
